+++ src/migrations/Install.php
@@ -38,7 +38,7 @@ class Install extends Migration
                 'fieldId' => $this->integer()->notNull(),
                 'typeId' => $this->integer()->notNull(),
                 'sortOrder' => $this->smallInteger()->unsigned(),
-                'level' => $this->smallInteger()->unsigned(),
+                // 'level' => $this->smallInteger()->unsigned(),
                 'deletedWithOwner' => $this->boolean()->null(),
                 'dateCreated' => $this->dateTime()->notNull(),
                 'dateUpdated' => $this->dateTime()->notNull(),